+++ src/integrator_trace.c
@@ -228,7 +228,7 @@ void reb_integrator_trace_dh_to_inertial(struct reb_simulation* r){
     temp.vz /= particles[0].m;
     // Use com to calculate central object's position.
     // This ignores previous values stored in particles[0].
-    // Should not matter unless collisions occured.
+    // Should not matter unless collisions occurred.
     particles[0].x = r->ri_trace.com_pos.x - temp.x;
     particles[0].y = r->ri_trace.com_pos.y - temp.y;
     particles[0].z = r->ri_trace.com_pos.z - temp.z;
@@ -471,7 +471,7 @@ void reb_integrator_trace_bs_step(struct reb_simulation* const r, double dt){
             star.vz = r->particles[0].vz;
 
             if (r->particles[0].x !=0 || r->particles[0].y !=0 || r->particles[0].z !=0){
-                // Collision with star occured
+                // Collision with star occurred
                 // Shift all particles back to heliocentric coordinates
                 // Ignore stars velocity:
                 //   - will not be used after this
@@ -593,7 +593,7 @@ void reb_integrator_trace_pre_ts_check(struct reb_simulation* const r){
 }
 
 double reb_integrator_trace_post_ts_check(struct reb_simulation* const r){
-    // This function returns 1 if any new encounters occured.
+    // This function returns 1 if any new encounters occurred.
     struct reb_integrator_trace* const ri_trace = &(r->ri_trace);
     const int N = r->N;
     const int Nactive = r->N_active==-1?r->N:r->N_active;
@@ -811,7 +811,7 @@ void reb_integrator_trace_step(struct reb_simulation* r){
     // Create copy of all particle to allow for the step to be rejected.
     memcpy(ri_trace->particles_backup, r->particles, N*sizeof(struct reb_particle));
 
-    // This will be set to 1 if a collision occured.
+    // This will be set to 1 if a collision occurred.
     ri_trace->force_accept = 0;
 
     // Check if there are any close encounters
@@ -820,7 +820,7 @@ void reb_integrator_trace_step(struct reb_simulation* r){
     // Attempt one step. 
     reb_integrator_trace_step_try(r);
 
-    // We alaways accept the step if a collision occured as it is impossible to undo the collision.
+    // We always accept the step if a collision occurred as it is impossible to undo the collision.
     if (!ri_trace->force_accept){
         // We check again for close encounters to ensure time reversibility. 
         if (reb_integrator_trace_post_ts_check(r)){